# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Image to text API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/get-vehicle-container": {
            "post": {
                "tags": [
                    "ImageToText"
                ],
                "summary": "This will return vehicle and container number",
                "operationId": "getVehicleContainerNumber",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "response_type"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file"
                                    },
                                    "response_type": {
                                        "type": "text",
                                        "enum": [
                                            "container",
                                            "vehicle"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}